requirements.txt
faiss-cpu==1.11.0
azure-identity==1.25.3
azure-search-documents==11.6.0
azure-storage-blob==12.28.0
ipykernel==6.29.5
ipython==8.32.0
notebook==7.3.2
langchain-community==0.4.1
langchain-core==1.2.20
langchain-openai==1.1.11
langchain-text-splitters==1.1.0
pypdf==6.4.2
python-dotenv==1.1.0
streamlit==1.45.1
